Prevent errors when updating operator pref summary
The response for manual network selection request may come after the
fragment that displays the search results is already destroyed, say
when the user moves out of the results screen. In this case, attempting
to update the summary for the chosen operator preference leads to an
exception.

Check whether the preference to be updated is still valid before
attempting to update its summary.

Bug: 161422555
Test: Manually tested - verified that no exception was thrown when the
response of manual network selection request was received after the
fragment was destroyed.
Change-Id: I087c0e101fa75f229bc4353defbb14e8bf30153f

Signed-off-by: Joey Huab <joey@evolution-x.org>
Signed-off-by: starlight5234 <starlight5234@protonmail.ch>